Refactor out all git IO utility functions into another module.
